chore: Update model config and training parameters
Adjust the model configuration and training parameters in order to improve performance and accuracy. Increase the decoder depth to 12 and the decoder heads to 16 in the model configuration. Modify the batch size to 2 in the train loader and set the number of workers to 1. Also, change the accumulate_grad_batches to 14 in the trainer configuration. These changes aim to optimize the training process and enhance the overall model performance.